+++ bin/install.sh
@@ -4,4 +4,4 @@ cd -P -- "$( dirname -- "${BASH_SOURCE[0]}" )/.."
 source ./bin/venv.sh
 
 echo "Installing..."
-pip install -qq -e ".[dev]"
+pip install -qq ".[dev]"